# Store evaluated value in object for future calls on Statamic/Fields/Value