The raw class variable is only set once on initialisation, so it makes sense to only evaluate the augmented value once too.

This PR adds a new class variable called value, which caches the value after being called once.

Before:
<img width="811" alt="Screenshot 2022-03-21 at 17 34 23" src="https://user-images.githubusercontent.com/25618897/159331904-cd04d0af-29fc-4f05-9e53-aeb2e6f65a24.png">

After:
<img width="810" alt="Screenshot 2022-03-21 at 17 34 30" src="https://user-images.githubusercontent.com/25618897/159331898-13e68d0b-8023-4fdd-aca7-d415c0b5d722.png">