@jasonvarga 2022-03-21
Nice!

Something I can see might happen is if an evaluated value ends up returning `null` or `false`, it will still try to evaluate it again each time.

Maybe you'll also need some sort of `hasEvaluated` property instead of just checking `value`.