@Daynnnnn 2022-04-11
As an extension to this, what are your thoughts on a cross request value cache? I understand the there's probably issues with the screenshot'd POC (invalidation logic, loads of cache save calls, and fieldtype duped handles being a few). But this has brought time spent on the value function down to 36ms

I'm not sure whether it's possible to cache further up, or at what point you'd rather people just use static caching 😅 Just thought it was worth noting what I'd found

<img width="534" alt="Screenshot 2022-04-11 at 14 48 20" src="https://user-images.githubusercontent.com/25618897/162754993-b9f210c6-ceb5-47d8-bd71-9e489de028bc.png">
<img width="783" alt="Screenshot 2022-04-11 at 14 45 09" src="https://user-images.githubusercontent.com/25618897/162755002-823c55d4-ef5d-48a3-af6d-c0087b2fb735.png">