@jasonvarga 2022-04-11
I think it's a cool idea, but let's leave it for a separate PR. I think it will need extra brainpower.
I'm thinking potential for race conditions, the cache value getting enormous, and more.

I remember we used to cache all the generated glide URLs to a single array. Eventually, as you generated more and more images, the array got so big that it started requiring more memory to unserialize it from the cache, and the site just crashed.

If we were to store all the augmented values in one big array, I could see the same problem. Potentially even sooner than the glide issue, since in there we were just storing URLs but in here we could be storing objects, etc.