@jasonvarga 2022-04-11
By the way, are you able to provide a GitHub repo for a site where this PR has a noticeable effect? If it's private, you can shoot it over to support@statamic.com. That would be great.