@simonhamp 2022-05-25
This is cool. I'm seeing around a ~100ms speed up on most page with this 👍🏼