@jacksleight 2022-10-19
Just been digging into a performance issue with a site and run into something that's kind of related to this, but not solved by this PR.

On this site (that uses Blade) I'm doing a lot of this:

```blade
{{ $entry->image ?? 'fallback.png' }}
```

What I've discovered is that is twice as slow as this:

```blade
{{ $entry->image }}
```
```php
// Only look this up once so fetch doesn't affect timings
$original = Entry::find('my-id');
Benchmark::dd([
    'get' => function () use ($original) {
        $entry = clone $original;
        $test = $entry->image;
    },
    'get-null-coalescing' => function () use ($original) {
        $entry = clone $original;
        $test = $entry->image ?? 'fallback.png';
    },
], 10);
```
```php
[
  "get" => "0.792ms"
  "get-null-coalescing" => "1.553ms"
]
```

I assume the reason is that the null coalescing causes the value to be evaluated twice.

However, this PR does not actually help in this situation because the evaluated value is stored in the `Value` object, but when you access an entry's value this way a brand new `Value` object is created every time. 

As a quick hack to test I added this to a custom `Entry` class:

```php
public function __get($key)
{
    return $this->evaluatedValues[$key] ?? ($this->evaluatedValues[$key] = parent::__get($key));
}
```

And now the score improves:

```php
[
  "get" => "0.789ms"
  "get-null-coalescing" => "0.777ms"
]
```

We're only talking a ms here, but on a page with lots of entries, values and null checks this could quickly add up.