@jasonvarga 2024-03-04
Closing this PR as it should be covered by the [upcoming performance improvements](https://github.com/statamic/cms/pull/9588). Thanks for understanding!